Fix test assertion: Corrected greeting from "Hello Juan" to "Hola Juan"
Corrected a mistake in the MockMvcTest.java file where the test expected the response to contain "Hello Juan" instead of "Hola Juan".

Since the application is in Spanish, the proper greeting should be "Hola Juan".

Changes made:

Updated MockMvcTest.java to expect "Hola Juan" instead of "Hello Juan".
Ensured that the test now aligns with the expected application output.
This fix prevents the test from failing due to an incorrect assertion.